#82 - WebCalendar month/year switch fix for last days of month #83 - Focus tracking issues fix in demo examples and various component Base StyleManager functionality added Base Skins (Themes) support added (only a few components are supported so far) Default WebLaF skin added (only a few components are supported so far) Dark WebLaF skin added (only a few components are supported so far) Alpha version of StyleEditor tool added Popup menu margin is now properly used within PopupMenuPainter to draw corner fill JScrollBarSettingsProcessor added for JScrollBar value auto-save possibility

+++ src/com/alee/extended/painter/InnerShadePainter.java
@@ -45,7 +45,7 @@ public class InnerShadePainter<E extends JComponent> extends NinePatchIconPainte
         super ();
     }
 
-    public InnerShadePainter ( final int shadeWidth, final int round, final int shadeOpacity )
+    public InnerShadePainter ( final int shadeWidth, final int round, final float shadeOpacity )
     {
         super ();
         this.shadeWidth = shadeWidth;
